Add the option EXPANSION_OPACITY_METHOD and apply the checksum-neutral fixes of the whole-codebase review (#621)
## What changed

Two commits on top of `main`:

1. **The option `EXPANSION_OPACITY_METHOD`.** The expansion opacity of a
wavelength bin sums (λ/Δλ) times a weight for each line in the bin. The
enum option selects the weight:

   | Value | Weight | Method |
   |---|---|---|
| `EXPANSION` | 1 − exp(−τ) | Eastman & Pinto 1993, ApJ, 412, 731–751,
doi:10.1086/172957 |
| `LINEBINNEDCAPPED` | min(1, τ) | line-binned opacity with a limit of 1
for each line |
| `LINEBINNED` | τ | line-binned opacity (Fontes, Fryer, Hungerford,
Wollaeger & Korobkin 2020, MNRAS, 493, 4143–4171,
doi:10.1093/mnras/staa485, their Eq. 9) |

`get_binned_opacity_line_weight()` holds the weight in one place. The
capped weight is `std::min(tau, 1.)`, which keeps a NaN, so the finite
check of the bin opacity still stops the run. `LINEBINNEDCAPPED` and
`LINEBINNED` need `RPKT_USE_EXPANSION_OPACITIES`, and a `static_assert`
stops other builds, because a line-by-line absorption has the weight 1 −
exp(−τ). The enum is in `constants.h`, and the option is in every preset
and in `artisoptions_doc.md`. All presets select `EXPANSION`, so the
option changes no result.

2. **The checksum-neutral fixes of a whole-codebase review
(2026-09-23).**

### Physics and packet transport

- The thermal emission with `RPKT_BOUNDBOUND_THERMALISATION_PROBABILITY
= P` weights the lines with P times the expansion opacity in the Planck
cumulative. Only the fraction P of a line absorption is a true
absorption, so the old weight broke Kirchhoff's law for P < 1. P = 1
(the CI value) gives the same numbers as before.
- A line absorption that thermalises gets the radiative energy factor of
the cell, as a free-free thermalisation does. The CI test has the factor
1.
- `update_pellet()` converts a pellet that decayed before `tmin` at the
first timestep of a new run, also when `timestep_initial` is above zero.
Before, such a run reached `__builtin_unreachable()`.
- The velocity grid of a virtual packet observer at -x uses a
right-handed basis.
- `get_nlte_vector_index()` counts the slots of a lower ion that has
autoionising levels above the NLTE cut and no superlevel. Before, the
slots of the next ion overlapped.
- A GPU build seeds each packet with `seed64()`. The old 32-bit mix gave
about 2.9e7 duplicate packet pairs for 5e8 packets.
- Static assertions reject `VPKT_ON` in a GPU build and
`DETAILED_LINE_ESTIMATORS_ON` with `RPKT_USE_EXPANSION_OPACITIES`.

### Robustness

- The nuclide mass rescale of a mapped model clamps a mass fraction at
1.
- A cell with a density and no element mass fractions stops the run.
- The gamma spectrum assertions of the standard nuclides now fire when a
table is absent. The XCOM reader checks each row. The seed line of
`input.txt` is checked.
- Twelve error sites use `fatal_crash()` in place of a log line and a
failed assertion.
- `initialise_linestat_file()` checks the stream after the write.
- The Makefile validates the ON/OFF options in one place (a value must
be exactly ON, OFF, or empty), keeps `assert_always()` active with an
inherited `-DNDEBUG`, escapes the `git status` text of `version.cc`, and
stops the compile database recipe when python fails.
- The test setup downloads the atomic data to a temporary file, so an
interrupted download leaves no truncated archive.

### Simplification

- Removed: the dead `DO_TITER` code and the `titer` parameters (about
120 lines; the `titeration` column of the estimators file stays),
`fracdep_ionisation_ion`, `n_auger_elec_avg`, `nautoionuptrans`, the
unreachable NLTE branch of `iongamma_is_zero()`, the duplicate-line
warning loop of `sort_temp_linelist()`, and the dead recovery code of
`boundary_distance()`.
- `emit_thermal_rpkt()` holds the thermal emission tail of the k-packet
code. `get_grey_tau_to_escape()` holds the ray integration of the
Wollaeger and Guttman schemes.
- `calculate_eff_ionpot_auger_rates()` returns the ionisation fraction,
so `analyse_sf_solution()` does not calculate the shell fractions again.
- `search_groundphixslist()` uses `upper_bound`.
- Each pass of `update_packets()` sorts only the active packets. The
multi-slot cell cache skips thick cells.
- The virtual packet counters are `ptrdiff_t`, and the contribution row
is built only with `VPKT_WRITE_CONTRIBS`.
- `exspec-after.sh` removes `vpackets_*.tmp`, the Leonardo scripts name
the correct job script, and the comments follow the STE rules.

## Checksums

The results of every CI test must stay identical. The P factor is 1 in
the only test with the thermalisation option, and every preset selects
`EXPANSION`.

## Left for separate pull requests (each changes results)

The review also found these defects. Each one changes the checksums of
one or more tests, so they are not in this pull request:

- `nltepop_write_to_file()` writes the -1 marker times the density for a
cell that took the thick path in the same update (the nebular tests at
timestep 4). The fix is a test of `elem_has_nlte_solution()`.
- At the last requested timestep `sn3d` writes the emission and
absorption files with `nts + 1` time blocks, and
`run-exspec-if-needed.sh` then skips `exspec`, so a run that stops
before the last timestep leaves files that artistools rejects.
- In LUT mode an element with zero mass fraction in a cell gives a
photoionisation estimator of exactly zero, and the excited levels of
other elements that map to that ground continuum get zero
photoionisation and bound-free heating.
- `call_T_e_finder()` stores the final T_e before the final residual
call, so the final ion balance of the elements without NLTE levels uses
the ionisation rates of the last trial temperature.
- A pellet without a gamma spectrum (52Fe, 52Mn) is counted as
non-thermal deposition but becomes a k-packet.
- `read_ion_transitions()` sorts on the non-unique key (lower, upper)
without `SORT_OR_STABLE_SORT`.
- The velocity grid time range of `vpkt.txt` is not checked against the
virtual packet time window. The grid range of `classicmode_1d_3dgrid`
lies outside its window, so its grid files are all zero.

## Verification

- `make OPTIMIZE=OFF sn3d exspec unittests` with clang: 182 of 182 unit
tests pass.
- A `TESTMODE=ON` build with `RPKT_USE_EXPANSION_OPACITIES` and the
thermalisation option compiles.
- `prek run --all-files` passes.


---------

+++ vpkt.h
@@ -3,6 +3,8 @@
 #ifndef VPKT_H
 #define VPKT_H
 
+#include <cstddef>
+
 #include "constants.h"
 #include "packet.h"
 
@@ -33,10 +35,10 @@ constexpr int VSPEC_TIMEBINS = 5;
 
 // separately cache-line aligned because they are incremented by all threads
 // per-rank counts of virtual packets created and escaped in the current timestep
-ALIGNAS_AVOID_FALSE_SHARING inline int nvpkt_created{0};
-ALIGNAS_AVOID_FALSE_SHARING inline int nvpkt_esc_from_rpkt{0};  // electron scattering event
-ALIGNAS_AVOID_FALSE_SHARING inline int nvpkt_esc_from_kpkt{0};  // thermal emission
-ALIGNAS_AVOID_FALSE_SHARING inline int nvpkt_esc_from_macroatom{0};  // macro-atom deactivation or line scattering
+ALIGNAS_AVOID_FALSE_SHARING inline ptrdiff_t nvpkt_created{0};
+ALIGNAS_AVOID_FALSE_SHARING inline ptrdiff_t nvpkt_esc_from_rpkt{0};  // electron scattering event
+ALIGNAS_AVOID_FALSE_SHARING inline ptrdiff_t nvpkt_esc_from_kpkt{0};  // thermal emission
+ALIGNAS_AVOID_FALSE_SHARING inline ptrdiff_t nvpkt_esc_from_macroatom{0};  // macro-atom deactivation or line scattering
 
 inline double optical_depth_is_thick_vpkt;
 }  // namespace vpkt
